requirements.txt
# Prediction Bot - Python Dependencies

# HTTP + async
requests>=2.31.0
aiohttp>=3.9.0

# Polymarket
py-clob-client>=0.18.0        # Official Polymarket CLOB SDK
web3>=6.15.0                   # Polygon/USDC on-chain execution

# Data + ML
pandas>=2.2.0
numpy>=1.26.0
xgboost>=2.0.0
scikit-learn>=1.4.0
scipy>=1.12.0

# Sentiment / NLP
praw>=7.7.0                    # Reddit API
tweepy>=4.14.0                 # Twitter API v2
feedparser>=6.0.11             # RSS feeds
vaderSentiment>=3.3.2          # Fast rule-based sentiment (baseline)
transformers>=4.38.0           # FinBERT for financial sentiment
torch>=2.2.0                   # Required by transformers

# Anthropic
anthropic>=0.25.0

# Storage
sqlalchemy>=2.0.0              # Optional ORM layer on top of SQLite

# Scheduling + utils
schedule>=1.2.0                # Simple cron-style scheduler
python-dotenv>=1.0.0           # Load .env credentials
loguru>=0.7.0                  # Better logging
rich>=13.7.0                   # Pretty CLI output
